+++ WorldCities/WorldCities/src/app/countries/countries.component.ts
@@ -3,6 +3,10 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatPaginator, PageEvent } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
+import { Subject } from 'rxjs';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+
+
 import { environment } from '../../envrionments/environment';
 import { Country } from './country';
 
@@ -25,6 +29,8 @@ export class CountriesComponent implements OnInit {
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
+
+  filterTextChanged: Subject<string> = new Subject<string>();
 
   constructor(private http: HttpClient) { }
 
@@ -63,6 +69,18 @@ export class CountriesComponent implements OnInit {
         this.paginator.pageSize = result.pageSize;
         this.countries = new MatTableDataSource<Country>(result.data);
       }, error => console.error(error));
+  }
+
+  // debounce filter text changes
+  onFilterTextChanged(filterText: string) {
+    if (this.filterTextChanged.observers.length == 0) {
+      this.filterTextChanged
+        .pipe(debounceTime(1000), distinctUntilChanged())
+        .subscribe(query => {
+          this.loadData(query);
+        });
+    }
+    this.filterTextChanged.next(filterText);
   }
 
 }